Menu update
It doesn't need the -1 as it's the length of the list not the position. So if the upgrade button is pressed once max level is reached the game no longer crashes.

However, support tower does not display max level as it only has two upgrades where as the other towers have 3.