Functionality of money
Called on the wrong attribute.
Now works perfectly

see screen shot

If the player doesn't have enough money to upgrade the tower the button simply doesn't do anything and the game continues. However it would be good to add something that indicates that the player does not have enough money.